(chore) Better naming and directory structure
