+++ src/TemplateStudio/forms/ModelForm.js
@@ -16,51 +16,47 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import InputErgo from '../presentational/InputErgo';
-import { Form, Segment, Tab, TextArea } from 'semantic-ui-react';
-import {UnControlled as ReactCodeMirror} from 'react-codemirror2';
-import CodeMirror from 'codemirror/lib/codemirror.js';
-import Resizable from 're-resizable';
-require('codemirror/addon/mode/simple.js');
+import ErgoInput from '../inputs/ErgoInput';
+import { Form, Segment, Tab } from 'semantic-ui-react';
 
-function showLogic(name) {
+function showModel(name) {
     return name.indexOf("@") == -1 && name.indexOf("system.cto") == -1;
 }
 
-class LogicForm extends Form {
+class ModelForm extends Form {
     constructor(props) {
         super(props);
-        this.handleLogicChange = this.handleLogicChange.bind(this);
+        this.handleModelChange = this.handleModelChange.bind(this);
         this.handleErgoMounted = this.handleErgoMounted.bind(this);
-        this.panesFromLogic = this.panesFromLogic.bind(this);
-    }
-
-    handleLogicChange(editor,name,logic) {
-        this.props.handleLogicChange(editor,name,logic);
+        this.panesFromModel = this.panesFromModel.bind(this);
     }
 
     handleErgoMounted(editor) {
         this.props.handleErgoMounted(editor);
     }
 
-    panesFromLogic(logic) {
+    handleModelChange(editor,name,model) {
+        this.props.handleModelChange(editor,name,model);
+    }
+
+    panesFromModel(model) {
         let panes = [];
-        for (const m of logic) {
-            if (showLogic(m.name))
+        for (const m of model) {
+            if (showModel(m.name))
                 panes.push({ 'menuItem': m.name, 'render': () =>
                              <Tab.Pane>
-                               <InputErgo
+                               <ErgoInput
                                  value={m.content}
                                  handleErgoMounted={(editor) => {this.handleErgoMounted(editor);}}
-                                 handleErgoChange={(editor,logic) => {this.handleLogicChange(editor,m.name,logic);}}/>
+                                 handleErgoChange={(editor,model) => {this.handleModelChange(editor,m.name,model);}}/>
                              </Tab.Pane> });
         }
         return panes;
     }
     
     render() {
-        const { logic, handleLogicChange } = this.props;
-        const panes = this.panesFromLogic(logic);
+        const { model, handleModelChange } = this.props;
+        const panes = this.panesFromModel(model);
         return (
             <Form>
               <Tab panes={panes} />
@@ -69,4 +65,4 @@ class LogicForm extends Form {
     }
 }
 
-export default LogicForm;
+export default ModelForm;
